The last commit: Added Ysm -> MyYsm converter
Unfortunately, I will not maintain this project anymore. The new version of ysm has scrambled all the packet ids and written the cache generation and some synchronization mechanisms into the jni layer. Although I know some of the logic, I still cannot start. The jni library part of ysm uses vmp and is virtualized, which makes it very difficult to work backwards. Therefore, it is not worthwhile and very energy-consuming for me to continue to do this task. Moreover, it may become more and more difficult to be compatible with ysm on different module loader platforms in the subsequent versions.
(Translated on https://fanyi.youdao.com/index.html with AI translation)